Make available the XPath path of the node where a SyntaxError occurred during Schema validation (#3316)
**What problem is this PR intended to solve?**

When validating an XML document with a schema it is difficult to
impossible to know where the error happened in the document. With the
path on the error, we can identify where in the document the error
occurred.

**Have you included adequate test coverage?**

I believe so. I updated the test in the schema to check to see if path
errors were being returned correctly.

**Does this change affect the behavior of either the C or the Java
implementations?**

This change the C implementation gently. Setting the path during error
creation.
I looked if xerces-j could do it, but I couldn't find anything on the
SAXException that would help.

---------

Co-authored-by: Mike Dalessio <mike.dalessio@gmail.com>