fix(css-selector): default XML namespace is not applied to wildcard
When performing a CSS selector query, an XML document's root namespace
declarations are injected into the XPathVisitor, and the default
namespace is applied to any non-namespaced element in the resulting
xpath query.

However, the wildcard selector "*" should not have this namespace
applied. This regressed in v1.17.0 with 179d74d5.

Fixes #3411