gumbo: Make sure to use the char* pointer as the hashmap item
Previously we were passing a character string of variable length, but
always copying 8 bytes into the item. It's not really a problem if the
length of the string is less than 8, but it's more serious if the
length is longer than 8.

Fixes https://github.com/sparklemotion/nokogiri/issues/3500
Fixes https://github.com/sparklemotion/nokogiri/issues/3508