CI: Remove ruby-3.1 from test matrix on Windows
Ruby-3.1 is not compatible to gcc-15 and patches will not be backported since it's EOL now.
Unfortunately MSYS2 is a rolling release with no option for an older gcc.
Althought setup-ruby installs an older gcc-14.2 package by default, it fails to install further packages per pacman due to dependencies:

```sh
  $ pacman.exe -Sy --noconfirm --noprogressbar --needed --disable-download-timeout  mingw-w64-ucrt-x86_64-libxml2 mingw-w64-ucrt-x86_64-libxslt
  resolving dependencies...
  looking for conflicting packages...
  error: failed to prepare transaction (could not satisfy dependencies)
  :: installing mingw-w64-ucrt-x86_64-gcc-libs (15.1.0-8) breaks dependency 'mingw-w64-ucrt-x86_64-gcc-libs=14.2.0-3' required by mingw-w64-ucrt-x86_64-gcc
```