Fix SIGINT during parsing crashing Ruby
We create HTML5 documents in a two step process designed to handle
exceptions being raised in the middle of the parsing process. There is a
race condition where a `SIGINT` received after creating the Ruby object
would cause the underlying `xmlDoc` to be freed in the `ensure` portion
of this process.

Fixes: #3528