dep: update to minitest 6
Note we have to jump through some new hoops to use bundler in
test-gem-install because ruby's bundled minitest is still v5 and we
want to be explicit about using v6.

Also note, we need to install build-essentials into the JRuby
containers to compile prism, which is a dependency of MT6.